refactor vtx image gen
